+++ lib/chat_web/live/main_live/page/login.ex
@@ -11,6 +11,7 @@ defmodule ChatWeb.MainLive.Page.Login do
   alias Chat.Identity
   alias Chat.Log
   alias Chat.User
+  alias Chat.UsersBroker
   alias ChatWeb.MainLive.Page
 
   @local_store_auth_key "buckitUp-chat-auth-v2"
@@ -19,13 +20,18 @@ defmodule ChatWeb.MainLive.Page.Login do
   def handshaked(socket), do: socket |> assign(:handshaked, true)
 
   def create_user(socket, name) do
-    me = User.login(name |> String.trim())
-    id = User.register(me)
+    me =
+      name
+      |> String.trim()
+      |> User.login()
+      |> tap(&User.register/1)
+      |> tap(&UsersBroker.put/1)
+
     # todo: check setting time before creating
     Log.sign_in(me, socket.assigns.monotonic_offset |> Chat.Time.monotonic_to_unix())
 
     socket
-    |> assign_logged_user(me, id)
+    |> assign_logged_user(me)
     |> store()
     |> close()
     |> Page.Lobby.notify_new_user(me |> Chat.Card.from_identity())
@@ -46,16 +52,16 @@ defmodule ChatWeb.MainLive.Page.Login do
   end
 
   def load_user(socket, %Identity{} = me, rooms) do
-    id =
-      me
-      |> User.login()
-      |> User.register()
+    me
+    |> User.login()
+    |> tap(&User.register/1)
+    |> tap(&UsersBroker.put/1)
 
     PubSub.subscribe(Chat.PubSub, login_topic(me))
     Log.visit(me, socket.assigns.monotonic_offset |> Chat.Time.monotonic_to_unix())
 
     socket
-    |> assign_logged_user(me, id, rooms)
+    |> assign_logged_user(me, rooms)
     |> close()
     |> Page.Lobby.notify_new_user(me |> Chat.Card.from_identity())
   end
@@ -149,10 +155,10 @@ defmodule ChatWeb.MainLive.Page.Login do
   defp login_topic(person),
     do: "login:" <> (person |> Enigma.hash() |> Base.encode16(case: :lower))
 
-  defp assign_logged_user(socket, me, id, rooms \\ []) do
+  defp assign_logged_user(socket, me, rooms \\ []) do
     socket
     |> assign(:me, me)
-    |> assign(:my_id, id)
+    |> assign(:my_id, Identity.pub_key(me))
     |> assign_rooms(rooms)
     |> maybe_create_admin_room()
   end