+++ CHANGELOG.md
@@ -11,11 +11,17 @@ Here we write upgrading notes for brands. It's a team effort to make them as
 straightforward as possible.
 
 ### Added
+- Search box for users/rooms
+- Users/Rooms sync optimization for UI
+- Integration of Secret Sharing with encryption layer (#251)
+- added shortcode for FE
 
 ### Changed
+- short code to be first 6 hex digits (was 8)
+
 
 ### Fixed
-
+- better broken files handling
 
 ## 2023-04-07_a41ea0d___2023-04-07_c06e36f
 
@@ -95,4 +101,7 @@ straightforward as possible.
 - UI: uploader covering chat/room list
 - lost feed names
 - admin room user actions
+
+
+